[BUG] occationally GP occurs when kill task. Add multilevel queue.
POC:
  createtask 2 1
  kill 2

+++ 02.Kernel64/Source/String.c
@@ -118,7 +118,7 @@ inline int toupper(int c) {
 
 char* strchr(const char* s, char c) {
   while (*s) {
-    char now = (int)*s;
+    char now = *s;
     if (now == c) {
       return (char*)s;
     }
